[test/dependency-injection] A testing ground
Some testing with a dependency-incjection system. Still have to
decide whether this is the way to go for this so this branch is
really ment for testing purpose only.
Please note that this branch might get a bunch of commit that
will never be used in the end product or gets trashed without
notice ;).